ml/requirements.txt
pandas
scikit-learn
psycopg2-binary
joblib